+++ Geral-core/src/br/cefetmg/inf/geral/model/domain/Parametro.java
@@ -2,8 +2,8 @@ package br.cefetmg.inf.geral.model.domain;
 
 public class Parametro {
     private String cod_Parametro;
-    private int qtd_Dia_Secagem;
-    private int qtd_Dia_Amamentacao;
+    private Long qtd_Dia_Secagem;
+    private Long qtd_Dia_Amamentacao;
 
     public String getCod_Parametro() {
         return cod_Parametro;
@@ -13,19 +13,19 @@ public class Parametro {
         this.cod_Parametro = cod_Parametro;
     }
 
-    public int getQtd_Dia_Secagem() {
+    public Long getQtd_Dia_Secagem() {
         return qtd_Dia_Secagem;
     }
 
-    public void setQtd_Dia_Secagem(int qtd_Dia_Secagem) {
+    public void setQtd_Dia_Secagem(Long qtd_Dia_Secagem) {
         this.qtd_Dia_Secagem = qtd_Dia_Secagem;
     }
 
-    public int getQtd_Dia_Amamentacao() {
+    public Long getQtd_Dia_Amamentacao() {
         return qtd_Dia_Amamentacao;
     }
 
-    public void setQtd_Dia_Amamentacao(int qtd_Dia_Amamentacao) {
+    public void setQtd_Dia_Amamentacao(Long qtd_Dia_Amamentacao) {
         this.qtd_Dia_Amamentacao = qtd_Dia_Amamentacao;
     } 
 }
